Split arithmetic tests out from test_operators
